+++ libraries/kotlinx-metadata/src/kotlinx/metadata/OptInMarkers.kt
@@ -1,0 +1,11 @@
+/*
+ * Copyright 2010-2022 JetBrains s.r.o. and Kotlin Programming Language contributors.
+ * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
+ */
+
+package kotlinx.metadata
+
+@RequiresOptIn(
+    "The API uses experimental feature \"context receivers\" (see KEEP-259) and may be changed or removed in any future release."
+)
+annotation class ExperimentalContextReceivers